Use int64 for DateEntered field
The `DateEntered` value expected to be a timestamp, so using int64
instead of uint make it easier to manipulate time.

+++ internal/monero/monero.go
@@ -53,7 +53,7 @@ type Node struct {
 	City            string         `json:"city" db:"city"`
 	Latitude        float64        `json:"latitude" db:"lat"`
 	Longitude       float64        `json:"longitude" db:"lon"`
-	DateEntered     uint           `json:"date_entered,omitempty" db:"date_entered"`
+	DateEntered     int64          `json:"date_entered,omitempty" db:"date_entered"`
 	LastChecked     int64          `json:"last_checked" db:"last_checked"`
 	FailedCount     uint           `json:"failed_count,omitempty" db:"failed_count"`
 	LastCheckStatus types.JSONText `json:"last_check_statuses" db:"last_check_status"`
